+++ src/sdk-exports.ts
@@ -130,22 +130,12 @@ export type {
   ApiError as DataClientApiError,
 } from "./types/data-client.types";
 export {
-  ACCESS_TOKEN_KEYS,
-  REFRESH_TOKEN_KEYS,
-  ACCESS_TOKEN_EXPIRES_AT_KEYS,
   normalizeExpiresAt,
   getJwtExpiresAt,
   getEffectiveUserTokenRefreshBuffer,
   getUserTokenRefreshDueAt,
   isUserTokenRefreshDue,
   isUserTokenExpired,
-  storeAccessToken,
-  storeRefreshToken,
-  clearStoredAccessToken,
-  clearStoredRefreshToken,
-  clearStoredSessionTokens,
-  getStoredRefreshToken,
-  getUserTokenExpiresAt,
   UserTokenRefreshManager,
 } from "./utils/user-token-refresh";
 export {